correction buf

+++ database/migrations/2020_02_01_141658_create_annoncereservations_table.php
@@ -20,7 +20,7 @@ class CreateAnnoncereservationsTable extends Migration
             $table->string('district')->nullable();
             $table->boolean('status')->default(true);
             $table->longText('description')->nullable();
-            $table->decimal('price', 15,2)->nullable();
+            $table->unsignedBigInteger('price')->nullable();
             $table->date('disponible_date')->nullable();
             $table->timestamps();
